requirements.txt
# Grant System Python Dependencies
# For Python 3.10+

# Environment & Configuration
python-dotenv>=1.0.0

# Web Scraping
selenium>=4.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Google Sheets Integration
gspread>=6.2.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Retry Logic & Resilience
tenacity>=8.2.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
flake8>=6.0.0
black>=23.0.0
